STY break round trip functions out into their own file
