+++ pyfolio/tests/test_txn.py
@@ -8,13 +8,10 @@ from pandas import (
     date_range,
     Timedelta
 )
-from pandas.util.testing import (assert_series_equal,
-                                 assert_frame_equal)
+from pandas.util.testing import (assert_series_equal)
 
 from pyfolio.txn import (get_turnover,
-                         adjust_returns_for_slippage,
-                         extract_round_trips,
-                         add_closing_transactions)
+                         adjust_returns_for_slippage)
 
 
 class TransactionsTestCase(TestCase):
@@ -87,64 +84,3 @@ class TransactionsTestCase(TestCase):
         result = adjust_returns_for_slippage(returns, turnover, slippage_bps)
 
         assert_series_equal(result, expected)
-
-    @parameterized.expand([
-        (DataFrame(data=[[2, 10, 'A'],
-                         [-2, 15, 'A']],
-                   columns=['amount', 'price', 'symbol'],
-                   index=dates[:2]),
-         DataFrame(data=[[dates[0], dates[1], Timedelta(days=1),
-                          10, True, 'A']],
-                   columns=['open_dt', 'close_dt',
-                            'duration', 'pnl', 'long', 'symbol'],
-                   index=[0])
-         ),
-        (DataFrame(data=[[2, 10, 'A'],
-                         [2, 15, 'A'],
-                         [-9, 10, 'A']],
-                   columns=['amount', 'price', 'symbol'],
-                   index=dates[:3]),
-         DataFrame(data=[[dates[0], dates[2], Timedelta(days=2),
-                          -10, True, 'A']],
-                   columns=['open_dt', 'close_dt',
-                            'duration', 'pnl', 'long', 'symbol'],
-                   index=[0])
-         ),
-        (DataFrame(data=[[2, 10, 'A'],
-                         [-4, 15, 'A'],
-                         [3, 20, 'A']],
-                   columns=['amount', 'price', 'symbol'],
-                   index=dates[:3]),
-         DataFrame(data=[[dates[0], dates[1], Timedelta(days=1),
-                          10, True, 'A'],
-                         [dates[1] + Timedelta(seconds=1), dates[2],
-                          Timedelta(days=1) - Timedelta(seconds=1),
-                          -10, False, 'A']],
-                   columns=['open_dt', 'close_dt',
-                            'duration', 'pnl', 'long', 'symbol'],
-                   index=[0, 1])
-         )
-    ])
-    def test_extract_round_trips(self, transactions, expected):
-        round_trips = extract_round_trips(transactions)
-        assert_frame_equal(round_trips, expected)
-
-    def test_add_closing_trades(self):
-        dates = date_range(start='2015-01-01', freq='D', periods=20)
-        transactions = DataFrame(data=[[2, 10, 'A'],
-                                       [-5, 10, 'A']],
-                                 columns=['amount', 'price', 'symbol'],
-                                 index=[dates[:2]])
-        positions = DataFrame(data=[[20, 0],
-                                    [-30, 30],
-                                    [-60, 30]],
-                              columns=['A', 'cash'],
-                              index=[dates[:3]])
-        expected = DataFrame(data=[[2, 10, 'A'],
-                                   [-5, 10, 'A'],
-                                   [3, 20., 'A']],
-                             columns=['amount', 'price', 'symbol'],
-                             index=[dates[:3]])
-
-        transactions_closed = add_closing_transactions(positions, transactions)
-        assert_frame_equal(transactions_closed, expected)
